test: given-when-then 구조로 변경
