+++ package.json
@@ -17,7 +17,7 @@
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "jest --config ./test/jest-e2e.json --testTimeout=60000 --runInBand comments.e2e-spec.ts"
+    "test:e2e": "jest --config ./test/jest-e2e.json --testTimeout=60000 --runInBand"
   },
   "dependencies": {
     "@nestjs/common": "^10.0.5",
